requirements.txt
Flask
Flask_SQLAlchemy
PyMySQL
markdown2
openai
python-dotenv
mysql-connector-python
Flask-CORS